requirements.txt
langchain==0.3.20
langchain_core==0.3.43
langchain_openai==0.1.16
langchain_community==0.3.19
langchain_ollama==0.2.3
langchain_groq==0.2.5
openai==1.13.3
pymupdf==1.25.3
pypdf==5.3.1
pydantic==2.10.6
python-dotenv==1.0.1